% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Shapfile World map}
\format{
A SpatialPolygonsDataFrame object for a world map
\describe{
}
}
\usage{
countries
}
\description{
Shapfile World map
}
\keyword{datasets}
