% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_WB}
\alias{data_WB}
\title{World Bank Data}
\format{
A data frame 10 variables:
\describe{
  \item{iso2c}{iso2c code}
  \item{iso3cl}{iso3c code}
  \item{country}{country code}
  \item{year}{}
  \item{life_expect}{life expectation}
  \item{population}{}
  \item{birth_rate}{}
  \item{neonat_mortal_rate}{}
  \item{region}{}
  \item{income}{}
}
}
\usage{
data_WB
}
\description{
World Bank Data
}
\keyword{datasets}
