% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{europ_countries}
\alias{europ_countries}
\title{Europ Countries}
\format{
A data frame about european countries:
\describe{
  \item{State}{}
  \item{Capital}{}
  \item{Population}{country code}
  \item{Area}{}
  \item{Density}{}
  \item{MEPs}{}
}
}
\usage{
europ_countries
}
\description{
Europ Countries
}
\keyword{datasets}
